+++ src/main/java/subway/application/LineService.java
@@ -4,8 +4,8 @@ import org.springframework.stereotype.Service;
 import subway.dao.LineDao;
 import subway.dao.entity.LineEntity;
 import subway.domain.Line;
-import subway.dto.LineRequest;
-import subway.dto.LineResponse;
+import subway.dto.request.LineRequest;
+import subway.dto.response.LineResponse;
 
 import java.util.List;
 import java.util.stream.Collectors;
