+++ src/main/java/subway/dto/request/ConnectRequest.java
@@ -1,4 +1,4 @@
-package subway.dto;
+package subway.dto.request;
 
 public class ConnectRequest {
     private final Long prevStationId;
